Successfully extract the filter component into a file. Now Filter.tsx contains both the KeywordSelect Component and the YearFilter component. Later we can add more functinoalities of filtering by editing the Filter.tsx
